Moved code into a "src" directory for cleanup
