+++ kits/daisy/src/robot/hexapod_parameters.cpp
