+++ kits/daisy/src/robot/hexapod_parameters.hpp
