+++ kits/daisy/src/xml_util/xml_helpers.cpp
